ai_assistant/requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
python-dotenv
openai
tqdm
scikit-learn
